move eg.py to bin/eg
We want to be able to run eg as a script, not as a runnable python
file. eg.py was running double duty as the script, and now it lives
as bin/eg instead. This means we don't have an immediate way to
invoke a .py file and run eg, but that seems like acceptable behavior
for the common case.