Page using less -R by default.
In a number of configurations, the default paging supported by the old
pydoc.pager() means of paging did not interpret ANSI color codes as
color, but instead showed raw characters. This change lets us use
`less -R` by default, which interprets color. It also adds the ability
to change the pager command string via a 'pager-cmd' option at the
command line and in the egrc. The new configuration tries to use
pydoc.pager in the event that less is not installed or if the
'pydoc.pager' flag is passed as the pager-cmd.