Rename modules to remove eg_* prefix.
The leading eg_* prefix was decided to be bad style and
redundant. eg_util is obviously the eg util, as you import it
from eg. Since we're breaking eg_exec.py with the next
version in order to support python 3.x, decided it makes
sense to throw in this other large change to further motivate
incrementing the minor version number.