Move du directory example to first section.
The example for using du to show the size of a directory
rather than listing all contents is a great one and seems so
useful that it should go in the first section. This moves
milouse's example to the first section. It also fixes a
spacing issue with find.md.