Add message for people using old symlink path.
Users that are invoking eg with a symlink rather than by
installing with eg are going to encounter issues when they
upgrade to eg 0.1.x, which had to push eg_exec.py up a level
in order to work with both python 2 and 3. This change adds a
helpful message if users invoke the old script, explaining
the change and briefly explaining how to fix it.