Remove checking to see if less is installed.
The way we were checking to see if less was installed before
paging with the default `less -R` command meant that user-
specified pager-cmd values beginning with `less` could fail,
even if less was installed. This was causing problems for
windows users. This commit removes this checking, meaning
that now `less -R` will always be the default value, even if
for some reason less isn't installed. In this case, users
should now either specify a custom pager-cmd or pass
pydoc.pager to let Python resolve a value. This could cause
problems for some users, but the thought is that less is
sufficiently ubiquitous that it won't be a problem for the
vast majority of users, and it is circumventable if necessary.