Update netcat to nc, add note on bsd vs gnu.
netcat and nc are interchangeable on different systems. I
settled on nc for more concise examples. The only basic case
that is differs on different systems, I think, is listening
either requiring or precluding the -p option. This commit
adds a second, implementation-specific example for this case
and adds a note on the difference between systems.