adb: add command to trim/clear all caches
Clearing all caches can be useful to get rid of e.g. broken files in cache where one does not know what app might be the culprit (if e.g. the device suddenly "behaves strangely" and it's not clear what causes it). So it would be good to have this command handy.